10-div.sql created and testde

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -6,7 +6,7 @@ DROP FUNCTION IF EXISTS SafeDiv;
 CREATE FUNCTION SafeDiv(a INT, b INT)
 RETURNS INT
 BEGIN
-    DECLARE answer INT;
+    DECLARE answer DECIMAL;
     IF b <> 0 THEN
         SET answer =  a / b;
     ELSE
